pyproject.toml
[tool.poetry]
name = "pysysml2"
version = "0.1.1"
description = "Python based parser for the SysML 2.0 textual modeling language."
authors = ["Keith Lucas <ke.le.luc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
anytree = "^2.8.0"
graphviz = "^0.20.1"
numpy = "^1.24.2"
pandas = "^2.0.1"
openpyxl = "^3.1.1"
antlr4-python3-runtime = "4.10"
typer = "^0.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
jupyter = "^1.0.0"
pytest-datadir = "^1.4.1"
poetry2setup = "^1.1.0"
poetry2conda = "^0.3.0"

[tool.poetry.scripts]
pysysml2 = "pysysml2.cli.__main__:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry_bumpversion.file."pysysml2/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[tool.poetry2conda]
name = "sysml2-env"

[tool.poetry2conda.dependencies]
anytree = { channel = "pip" }
graphviz = { channel = "pip" }
antlr4-python3-runtime = { channel = "pip" }
